+++ app/interest_zone.py
@@ -1,0 +1,6 @@
+class InterestZone() :
+    def __init__(self, id, color_name, color, rectangle) :
+        self.id = id
+        self.color_name = color_name
+        self.color = color
+        self.rectangle = rectangle
