+++ app/status.py
@@ -1,0 +1,5 @@
+from enum import StrEnum
+
+class Status:
+    WALKING = "Walking"
+    STANDING = "Standing"
