--- status.py
@@ -1,5 +1,0 @@
-from enum import StrEnum
-
-class Status:
-    WALKING = "Walking"
-    STANDING = "Standing"
